requirements.txt
bcrypt==4.1.2
boto3==1.38.11
fastapi==0.115.12
mysql-connector-python==8.3.0
pandas==2.2.3
pydantic==2.11.5
PyJWT==2.10.1
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
twilio==9.6.1
uvicorn[standard]==0.34.0
python-multipart==0.0.9
